magento2-login-as-customer/issues/104: Global refactoring
- Extract LoginAsCustomerApi
