+++ app/code/Magento/LoginAsCustomerApi/Api/ConfigInterface.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\LoginAsCustomer\Api;
+namespace Magento\LoginAsCustomerApi\Api;
 
 /**
  * LoginAsCustomer config
