+++ app/code/Magento/LoginAsCustomerApi/README.md
@@ -1,0 +1,3 @@
+# Magento_LoginAsCustomer module
+
+The Magento_LoginAsCustomerApi module provides API for ability to login into customer account for an admin user.
