+++ app/code/Magento/LoginAsCustomerApi/composer.json
@@ -1,0 +1,19 @@
+{
+    "name": "magento/module-login-as-customer-api",
+    "description": "Allow for admin to enter a customer account",
+    "require": {
+        "php": "~7.1.3||~7.2.0||~7.3.0",
+        "magento/framework": "*"
+    },
+    "type": "magento2-module",
+    "license": [
+        "OSL-3.0",
+        "AFL-3.0"
+    ],
+    "autoload": {
+        "files": [ "registration.php" ],
+        "psr-4": {
+            "Magento\\LoginAsCustomerApi\\": ""
+        }
+    }
+}
