+++ app/code/Magento/LoginAsCustomerSales/etc/di.xml
@@ -7,7 +7,7 @@
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="Magento\LoginAsCustomer\Api\AuthenticateCustomerInterface">
+    <type name="Magento\LoginAsCustomerApi\Api\AuthenticateCustomerInterface">
         <plugin name="login_as_customer_sales_authenticate_customer"
                 type="Magento\LoginAsCustomerSales\Plugin\AuthenticateCustomer"/>
     </type>
