+++ app/code/Magento/LoginAsCustomerUi/composer.json
@@ -4,7 +4,7 @@
     "require": {
         "php": "~7.1.3||~7.2.0||~7.3.0",
         "magento/framework": "*",
-        "magento/module-login-as-customer": "*",
+        "magento/module-login-as-customer-api": "*",
         "magento/module-backend": "*",
         "magento/module-customer": "*",
         "magento/module-sales": "*",
